requirements-dev.txt
-r requirements.txt
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0
isort==5.12.0
pre-commit==3.4.0